+++ sam/metrics/__init__.py
@@ -3,10 +3,11 @@ from .incident_recall import incident_recall, make_incident_recall_scorer,\
 from .mase import mean_absolute_scaled_error
 from .keras_metrics import keras_tilted_loss, keras_joint_mse_tilted_loss, \
     keras_rmse, get_keras_forecasting_metrics
+from .train_mean_r2 import train_mean_r2
 
 from . import mase
 from . import keras_metrics
 
 __all__ = ["incident_recall", "make_incident_recall_scorer", "precision_incident_recall_curve",
            "mean_absolute_scaled_error",
-           "keras_tilted_loss", "keras_rmse", "get_keras_forecasting_metrics"]
+           "keras_tilted_loss", "keras_rmse", "get_keras_forecasting_metrics", "train_mean_r2"]