+++ sam/visualization/__init__.py
@@ -7,6 +7,7 @@ from .deprecated import make_precision_recall_curve, make_threshold_plot, \
     make_incident_heatmap
 from .extreme_removal_plot import diagnostic_extreme_removal
 from .quantile_plot import sam_quantile_plot
+from .performance_evaluation_fixed_predict_ahead import performance_evaluation_fixed_predict_ahead
 
 from . import precision_recall
 from . import threshold_plot
@@ -14,4 +15,5 @@ from . import incident_heatmap
 from . import rolling_correlations
 
 __all__ = ["plot_precision_recall_curve", "plot_threshold_curve", "plot_incident_heatmap",
-           "plot_lag_correlation", "diagnostic_extreme_removal"]
+           "plot_lag_correlation", "diagnostic_extreme_removal", "sam_quantile_plot",
+           "performance_evaluation_fixed_predict_ahead"]